+++ Adj/Program/Casting.hs
@@ -1,11 +1,11 @@
-module Adj.Program.Disbandable where
+module Adj.Program.Casting where
 
 import Adj.Algebra (Flat (Flat), Dual (Dual), Kleisli (Kleisli))
 
 infixl 7 =-=, -=-
 infixl 8 =-, -=
 
-class Disbandable t where
+class Casting t where
 	{-# MINIMAL (=-), (-=) #-}
 	type Primary t a :: *
 	(=-) :: t a -> Primary t a
@@ -17,17 +17,17 @@ class Disbandable t where
 	(-=-) :: (t a -> t b) -> Primary t a -> Primary t b
 	m -=- x = (=-) (m ((-=) x))
 
-instance Disbandable (Flat morphism source) where
+instance Casting (Flat morphism source) where
 	type Primary (Flat morphism source) target = morphism source target
 	(=-) (Flat m) = m
 	(-=) m = Flat m
 
-instance Disbandable (Dual morphism target) where
+instance Casting (Dual morphism target) where
 	type Primary (Dual morphism target) source = morphism source target
 	(=-) (Dual m) = m
 	(-=) m = Dual m
 
-instance Disbandable (Kleisli effect morphism source) where
+instance Casting (Kleisli effect morphism source) where
 	type Primary (Kleisli effect morphism source) target = morphism source (effect target)
 	(=-) (Kleisli m) = m
 	(-=) m = Kleisli m