+++ Adj/Program/Construction.hs
@@ -1,11 +1,11 @@
 module Adj.Program.Construction where
 
 import Adj.Algebra ((.:), Functor (map), (|-|->), (:*:) ((:*:)), type (-->), Flat (Flat))
-import Adj.Program.Disbandable (Disbandable (Primary, (=-), (-=)))
+import Adj.Program.Casting (Casting (Primary, (=-), (-=)))
 
 newtype Construction t a = Construction (a :*: t (Construction t a))
 
-instance Disbandable (Construction t) where
+instance Casting (Construction t) where
 	type Primary (Construction t) a = a :*: t (Construction t a)
 	(=-) (Construction m) = m
 	(-=) m = Construction m