+++ Adj/Program/Instruction.hs
@@ -1,11 +1,11 @@
 module Adj.Program.Instruction where
 
 import Adj.Algebra.Category ((.:), Functor (map), (|-|->), (:+:) (Option, Adoption), type (-->), Flat (Flat))
-import Adj.Program.Disbandable (Disbandable (Primary, (=-), (-=)))
+import Adj.Program.Casting (Casting (Primary, (=-), (-=)))
 
 newtype Instruction t a = Instruction (a :+: t (Instruction t a))
 
-instance Disbandable (Instruction t) where
+instance Casting (Instruction t) where
 	type Primary (Instruction t) a = a :+: t (Instruction t a)
 	(=-) (Instruction m) = m
 	(-=) m = Instruction m